feat: Create universal file upload zod string creator
